Add the Environment ID to the influxdb data.
Also move the interface of an execution environment into its own file, execution_environment.go.
