+++ internal/api/api.go
@@ -47,8 +47,8 @@ func configureV1Router(router *mux.Router,
 		w.WriteHeader(http.StatusNotFound)
 	})
 	v1 := router.PathPrefix(BasePath).Subrouter()
-	v1.HandleFunc(HealthPath, Health).Methods(http.MethodGet)
-	v1.HandleFunc(VersionPath, Version).Methods(http.MethodGet)
+	v1.HandleFunc(HealthPath, Health).Methods(http.MethodGet).Name(HealthPath)
+	v1.HandleFunc(VersionPath, Version).Methods(http.MethodGet).Name(VersionPath)
 
 	runnerController := &RunnerController{manager: runnerManager}
 	environmentController := &EnvironmentController{manager: environmentManager}
@@ -59,7 +59,8 @@ func configureV1Router(router *mux.Router,
 		// May add a statistics controller if another route joins
 		statisticsRouter := router.PathPrefix(StatisticsPath).Subrouter()
 		statisticsRouter.
-			HandleFunc(EnvironmentsPath, StatisticsExecutionEnvironments(environmentManager)).Methods(http.MethodGet)
+			HandleFunc(EnvironmentsPath, StatisticsExecutionEnvironments(environmentManager)).
+			Methods(http.MethodGet).Name(EnvironmentsPath)
 	}
 
 	if auth.InitializeAuthentication() {
