+++ internal/api/websocket.go
@@ -8,7 +8,6 @@ import (
 	"github.com/gorilla/websocket"
 	"github.com/openHPI/poseidon/internal/runner"
 	"github.com/openHPI/poseidon/pkg/dto"
-	"github.com/openHPI/poseidon/pkg/logging"
 	"io"
 	"net/http"
 	"sync"
@@ -334,7 +333,8 @@ func (wp *webSocketProxy) writeMessage(messageType int, data []byte) error {
 // connectToRunner is the endpoint for websocket connections.
 func (r *RunnerController) connectToRunner(writer http.ResponseWriter, request *http.Request) {
 	targetRunner, _ := runner.FromContext(request.Context())
-	logging.AddRunnerID(request, targetRunner.ID())
+	addMonitoringData(request, targetRunner)
+
 	executionID := request.URL.Query().Get(ExecutionIDKey)
 	if !targetRunner.ExecutionExists(executionID) {
 		writeNotFound(writer, ErrUnknownExecutionID)
